+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Added getClientIp method to Request.
 - Added getReferrer method to Request.
+- Added getCookieValue method to Request.
 - Added ErrorControllerTrait.
 - Added isPreActionEventEnabled and isPostActionEventEnabled methods to Controller.
 - Prevented onPreActionEvent and onPreActionEvent methods from running on error controller if an exception occurred.